Split tray menu sections into files
